Prioritized Scheduling Routing Protocol
The development of wireless body area networks, or WBANs, has altered people's lives through their utilization in the fields of athletics, cultural activities, fitness, including healthcare, among others. Energy conservation and ensuring the quality of offerings, however, are two of the main design difficulties for WBAN. In a WBAN, the load balancing of various packet buffers is crucial to the construction of a dependable and environmentally friendly technology. This paper proposes a prioritized scheduling-based protocol for minimizing packet drops in wireless body area networks on IEEE 802.15.6. This paper's primary goal is to reduce packet drops in the queues to increase WBAN throughput. In this instance, we take into account the data packet's importance as well as its source location to ensure that no packet is held in the designated buffer for an extended period before being sent to the connection point. PyCrypto is used to replicate the suggested approach in order to research and contrast its results with those of its competitors. According to the findings from the simulation, the suggested protocol performs more efficiently in delay, throughput, and energy consumption than the current approaches.